fix: support auth on web
